requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
Django==3.0.5
django-cors-headers==3.6.0
django-mysql==3.10.0
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
django-toolbox==0.2
djangorestframework==3.12.2
djangotoolbox==1.8.0
djongo==1.3.3
et-xmlfile==1.0.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jsonfield==3.1.0
lxml==4.6.2
MarkupSafe==1.1.1
mongoengine ==0.22.0
numpy     ==                        1.19.4
openpyxl      ==                    3.0.5
pandas   ==                         1.1.5
Pillow     ==                       8.0.1
psycopg2   ==                       2.8.6
pymongo    ==                       3.11.2
python-dateutil  ==                 2.8.1
python-pptx     ==                  0.6.18
pytz      ==                        2020.4
requests  ==                        2.25.1
selenium  ==                        3.141.0
six     ==                          1.15.0
soupsieve ==                        2.1
sqlparse  ==                        0.2.4
urllib3     ==                      1.26.2
webdriver-manager  ==               3.2.2
Werkzeug         ==                 1.0.1
xlrd             ==                 2.0.1
XlsxWriter       ==                 1.3.7
gunicorn == 20.0.4